+++ resources/views/partials/footer.blade.php
@@ -1,7 +1,0 @@
-<footer class="content-info">
-  <div class="container">
-    @php dynamic_sidebar('sidebar-footer') @endphp
-  </div>
-</footer>
-
-@includeWhen(!$is_production, 'tools.guide')