+++ vite.config.ts
@@ -1,13 +1,15 @@
 import dotenv from 'dotenv'
 import path from 'path'
 import { defineConfig, ServerOptions } from 'vite'
+import outputManifest, { Bundle, KeyValueDecorator } from 'rollup-plugin-output-manifest'
+import copy from 'rollup-plugin-copy'
 
-dotenv.config()
+dotenv.config({ path: '../../../../.env' })
 
 const assets = {
-    base: 'resources/assets',
-    scripts: 'resources/assets/scripts',
-    styles: 'resources/assets/styles'
+    base: 'resources',
+    scripts: 'resources/scripts',
+    styles: 'resources/styles'
 }
 
 const server: ServerOptions = {
@@ -27,6 +29,7 @@ process.env.HTTPS_KEY &&
     })
 
 export default defineConfig({
+    publicDir: false,
     resolve: {
         alias: {
             '@': path.resolve(__dirname, assets.base),
@@ -35,13 +38,76 @@ export default defineConfig({
         }
     },
     build: {
-        sourcemap: true,
-        manifest: true,
+        sourcemap: 'inline',
+        manifest: false,
+        outDir: 'public',
+        assetsDir: '',
         rollupOptions: {
-            input: [`${assets.scripts}/main.ts`, `${assets.styles}/main.scss`],
+            input: {
+                main: path.resolve(__dirname, `${assets.scripts}/main.ts`),
+                editor: path.resolve(__dirname, `${assets.scripts}/editor.ts`)
+            },
             output: {
-                dir: 'dist'
-            }
+                sourcemap: true
+            },
+            plugins: [
+                outputManifest({
+                    fileName: 'manifest.json',
+                    map: (bundle: Bundle): Bundle => {
+                        bundle.name = bundle.name.replace(/.css$/gm, '')
+
+                        return bundle
+                    }
+                }),
+                outputManifest({
+                    fileName: 'entrypoints.json',
+                    map: (bundle: Bundle): Bundle => {
+                        bundle.name = bundle.name.replace(/.css$/gm, '')
+
+                        return bundle
+                    },
+                    generate: (_: KeyValueDecorator, seed: object) => chunks =>
+                        chunks.reduce((manifest, { name, fileName }) => {
+                            const output = {}
+                            const js = manifest[name] ? manifest[name].js : []
+                            const css = manifest[name] ? manifest[name].css : []
+                            const dependencies = manifest[name] ? manifest[name].dependencies : []
+
+                            if (fileName.match(/.js$/gm)) {
+                                js.push(fileName)
+                            }
+
+                            if (fileName.match(/.css$/gm)) {
+                                css.push(fileName)
+                            }
+
+                            output[name] = { js, css, dependencies }
+
+                            return {
+                                ...manifest,
+                                ...output
+                            }
+                        }, seed)
+                }),
+                copy({
+                    copyOnce: true,
+                    hook: 'writeBundle',
+                    targets: [
+                        {
+                            src: path.resolve(__dirname, `${assets.base}/images/**/*`),
+                            dest: 'public/images'
+                        },
+                        {
+                            src: path.resolve(__dirname, `${assets.base}/svg/**/*`),
+                            dest: 'public/svg'
+                        },
+                        {
+                            src: path.resolve(__dirname, `${assets.base}/fonts/**/*`),
+                            dest: 'public/fonts'
+                        }
+                    ]
+                })
+            ]
         }
     },
     server