+++ CHANGELOG.md
@@ -1,7 +1,20 @@
-# CHANGELOG
-
-## v2.2.0
+## [3.0.0](///compare/2.2.0...3.0.0) (2022-03-25)
 
 ### Features
 
--   Modular stylesheet architecture
+-   add bundler copy plugin 4a3e9c4
+-   add helpers 5e1ff1d
+-   add vite-env type reference 903934d
+-   dev guide 5c71091
+-   enable hot reload 6b1fd28
+-   fork of sage 10 0299e4a
+-   import helpers b295269
+-   set up vite 9887ad1
+-   setup phpcs bcb6ce9
+-   update screenshot 7325c68
+
+## 2.2.0
+
+### Features
+
+-   modular stylesheet architecture